requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
deap==1.4.2
debugpy==1.8.9
decorator==5.1.1
distlib==0.3.9
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-paginate==2024.4.12
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nx==0.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.8.0
pyparsing==3.1.2
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
python-telegram-bot==21.4
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.24.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
SQLAlchemy==2.0.36
sqlparse==0.5.0
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.27.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.17.0
wsproto==1.2.0
XlsxWriter==3.2.0
